requirements.txt
elasticsearch==8.14.0
pandas
websockets
pyais
python-dotenv
requests